add 3 small components, helper function
